fix(server): cast event.reminder_lead_minutes to BIGINT + sync make_interval bigint
Two PG-16 type-mismatch fixes:

1. server/src/handlers/event.rs — `event.reminder_lead_minutes` is INT4 in
   the DB but the Rust `Event` struct uses `Option<i64>`. EVENT_SELECT
   already had a `::BIGINT` cast for the read path; the update handler's
   pre-flight SELECT did not, so PUT /api/events/:id 500'd on every edit.
2. server/src/handlers/sync.rs — `make_interval` only has `integer` and
   multi-arg overloads; binding `i64` produced
   `function make_interval(days => bigint) does not exist` and the hourly
   prune loop error-spammed the log. Cast `days => $1::int`.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)

+++ server/src/handlers/sync.rs
@@ -455,9 +455,11 @@ pub async fn pull(
 }
 
 pub async fn prune_change_log(pool: &PgPool, ttl_days: i64) -> Result<u64, sqlx::Error> {
+    // `make_interval` only takes `integer`, but sqlx binds `i64` → bigint, so we
+    // cast explicitly. PG 16 has no overload for bigint days.
     let result = sqlx::query(
         "DELETE FROM sync_change_log \
-         WHERE changed_at < to_char(NOW() AT TIME ZONE 'UTC' - make_interval(days => $1), 'YYYY-MM-DD HH24:MI:SS')"
+         WHERE changed_at < to_char(NOW() AT TIME ZONE 'UTC' - make_interval(days => $1::int), 'YYYY-MM-DD HH24:MI:SS')"
     )
     .bind(ttl_days)
     .execute(pool)
